# perf: do not lowercase returned values