@lucasfcosta 2016-10-08
LGTM! I agree with that, it should be a user's responsibility to sanitize the output of this module.

Also, when committing a breaking change you just gotta add `BREAKING CHANGE: <description>` to the commit right?