@keithamus 2016-10-08
@lucasfcosta yup indeed. I've added that to the commit, quite far down. IMO it's not the most ideal as its easy to forget/mess up and release a patch/feature version instead. I suppose we'll just have to now start scrutinising commit messages too.