@meeber 2016-10-08
At dinner but will review in about an hour :D